+++ artifacts/kuhni-na-zakaz/.next/static/chunks/webpack.js
@@ -190,7 +190,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-/******/ 		__webpack_require__.h = () => ("0dd03f74ecaf1a8a")
+/******/ 		__webpack_require__.h = () => ("796f26dde552037f")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/global */